merging

+++ src/execution/player.c
@@ -6,7 +6,7 @@
 /*   By: rkaras <rkaras@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/16 18:47:44 by rkaras        #+#    #+#                 */
-/*   Updated: 2025/03/27 12:59:51 by rkaras        ########   odam.nl         */
+/*   Updated: 2025/03/27 14:46:40 by sramos        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,9 +55,9 @@ void	move_player(t_data *data, double move_x, double move_y)
 	if (map_x < 0 || map_y < 0 || map_x >= data->map_width
 		|| map_y >= data->map_height)
 		return ;
-	if (data->map[(int)map_y][(int)map_x] != '1' &&
-		data->map[(int)map_y][data->player->pos_x / TILE_SIZE] != '1' &&
-		data->map[data->player->pos_y / TILE_SIZE][(int)map_x] != '1')
+	if (data->map[map_y][map_x] != '1' &&
+		data->map[map_y][data->player->pos_x / TILE_SIZE] != '1' &&
+		data->map[data->player->pos_y / TILE_SIZE][map_x] != '1')
 	{
 		data->player->pos_x = new_x;
 		data->player->pos_y = new_y;
